backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Supabase
supabase==2.7.4
httpx==0.27.0

# ML — core
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# ML — NLP
sentence-transformers==3.1.1
nltk==3.9.1

# Agentic AI
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
anthropic==0.34.2

# Integrations
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
slack-sdk==3.31.0
jira==3.8.0
requests==2.32.3

# Pydantic
pydantic==2.9.2
tqdm==4.66.5